[migrate] - tauri 2 automatic migration
