+++ src/Syncer.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useReducer } from "react";
-import { invoke } from "@tauri-apps/api/tauri";
+import { invoke } from "@tauri-apps/api/core";
 
 import type { AppState } from "./types";
